+++ lib/src/Message.h
@@ -53,7 +53,7 @@ class Message: public WireFormatter, public PacketWriter
     long long offset;
 
     Message(Packet *packet, long long offset = -1);
-    Message(int crc, unsigned char magicByte, unsigned char attributes, int keyLength, unsigned char* key, int valueLength, unsigned char* value, long int offset = -1, bool releaseArrays = false);
+    Message(int crc, unsigned char magicByte, unsigned char attributes, int keyLength, unsigned char* key, int valueLength, unsigned char* value, long long offset = -1, bool releaseArrays = false);
     ~Message();
 
     unsigned char* toWireFormat(bool updatePacketSize = true);
